pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Panda" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
requires-python = "~=3.12.0"


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true


[tool.poetry.dependencies]
supabase = "^2.17.0"
openai = "^1.97.0"
pandas = "^2.3.1"
bokeh = "^3.7.3"
tqdm = "^4.67.1"
dotenv = "^0.9.9"
loguru = "^0.7.3"
beautifulsoup4 = "^4.13.4"
requests = "^2.32.4"
lxml = "^6.0.0"
pytest = "^8.4.1"
google-genai = "^1.26.0"
typer = "^0.16.0"
